+++ src/Perpetuum/Perpetuum.csproj
@@ -717,8 +717,12 @@
     <Compile Include="Zones\NpcSystem\NpcBossInfo.cs" />
     <Compile Include="Zones\NpcSystem\NpcEp.cs" />
     <Compile Include="Zones\NpcSystem\Presences\ExpiringStaticPresence\FlockEscalationSelector.cs" />
+    <Compile Include="Zones\NpcSystem\Presences\ExpiringStaticPresence\IRandomStaticPresence.cs" />
     <Compile Include="Zones\NpcSystem\Presences\ExpiringStaticPresence\StaticExpiringFlock.cs" />
     <Compile Include="Zones\NpcSystem\Presences\ExpiringStaticPresence\RandomSpawningExpiringPresence.cs" />
+    <Compile Include="Zones\NpcSystem\Presences\ExpiringStaticPresence\StaticSpawnState.cs" />
+    <Compile Include="Zones\NpcSystem\Presences\GrowingPresences\GrowingPresence.cs" />
+    <Compile Include="Zones\NpcSystem\Presences\GrowingPresences\GrowthStates.cs" />
     <Compile Include="Zones\NpcSystem\Reinforcements\INpcReinforcementWave.cs" />
     <Compile Include="Zones\NpcSystem\Reinforcements\INpcReinforcementsRepository.cs" />
     <Compile Include="Zones\NpcSystem\Reinforcements\INpcReinforcements.cs" />
