+++ src/Perpetuum/Zones/NpcSystem/Presences/ExpiringStaticPresence/IRandomStaticPresence.cs
@@ -1,0 +1,4 @@
+﻿namespace Perpetuum.Zones.NpcSystem.Presences.ExpiringStaticPresence
+{
+    interface IRandomStaticPresence : IRoamingPresence { }
+}
