+++ src/Perpetuum/Zones/NpcSystem/Presences/ExpiringStaticPresence/RandomSpawningExpiringPresence.cs
@@ -1,13 +1,7 @@
 ﻿using Perpetuum.StateMachines;
 using Perpetuum.Zones.NpcSystem.Presences.PathFinders;
 using System;
-using System.Linq;
 using System.Drawing;
-using Perpetuum.ExportedTypes;
-using Perpetuum.Units.DockingBases;
-using Perpetuum.Units;
-using Perpetuum.Zones.Teleporting;
-using Perpetuum.Timers;
 using Perpetuum.Zones.NpcSystem.Presences.ExpiringStaticPresence;
 
 namespace Perpetuum.Zones.NpcSystem.Presences.RandomExpiringPresence
@@ -15,7 +9,7 @@ namespace Perpetuum.Zones.NpcSystem.Presences.RandomExpiringPresence
     /// <summary>
     /// A non-roaming ExpiringPresence that would spawning with Roaming rules
     /// </summary>
-    public class RandomSpawningExpiringPresence : ExpiringPresence, IRoamingPresence
+    public class RandomSpawningExpiringPresence : ExpiringPresence, IRandomStaticPresence, IRoamingPresence
     {
         public StackFSM StackFSM { get; protected set; }
         public Position SpawnOrigin { get; set; }
@@ -58,204 +52,6 @@ namespace Perpetuum.Zones.NpcSystem.Presences.RandomExpiringPresence
         public void OnSpawned()
         {
             ResetDynamicDespawnTimer();
-        }
-    }
-
-    public class GrowingPresence : RandomSpawningExpiringPresence
-    {
-        public TimeSpan GrowTime { get; private set; }
-        public IEscalatingPresenceFlockSelector Selector { get; private set; }
-        public int CurrentGrowthLevel { get; private set; }
-        public GrowingPresence(IZone zone, IPresenceConfiguration configuration, IEscalatingPresenceFlockSelector selector) : base(zone, configuration)
-        {
-            Selector = selector;
-            if (Configuration.GrowthSeconds != null)
-                GrowTime = TimeSpan.FromSeconds((int)Configuration.GrowthSeconds);
-        }
-
-        protected override void InitStateMachine()
-        {
-            CurrentGrowthLevel = FastRandom.NextInt(9);
-            StackFSM = new StackFSM();
-            StackFSM.Push(new NPCBaseGrowState(this));
-        }
-
-        public override void LoadFlocks()
-        {
-            for (var i = 0; i <= CurrentGrowthLevel; i++)
-            {
-                var flockConfigs = Selector.GetFlocksForPresenceLevel(this, i);
-                foreach (var config in flockConfigs)
-                {
-                    CreateAndAddFlock(config);
-                }
-            }
-        }
-
-        public void OnWaveSpawn(int level)
-        {
-            CurrentGrowthLevel = Math.Max(CurrentGrowthLevel, level);
-        }
-
-        protected override void OnPresenceExpired()
-        {
-            base.OnPresenceExpired();
-            ClearFlocks();
-            CurrentGrowthLevel = 0;
-            LoadFlocks();
-        }
-    }
-
-    public class StaticSpawnState : SpawnState
-    {
-        private readonly int BASE_RADIUS = 300;
-        private readonly int PLAYER_RADIUS = 150;
-        public StaticSpawnState(IRoamingPresence presence, int playerMinDist = 200) : base(presence, playerMinDist) { }
-
-        protected override void OnSpawned()
-        {
-            _presence.OnSpawned();
-            _presence.StackFSM.Push(new NullRoamingState(_presence));
-        }
-
-        protected override bool IsInRange(Position position, int range)
-        {
-            var zone = _presence.Zone;
-            if (zone.Configuration.IsGamma && zone.IsUnitWithCategoryInRange(CategoryFlags.cf_pbs_docking_base, position, BASE_RADIUS))
-                return true;
-            else if (zone.GetStaticUnits().OfType<DockingBase>().WithinRange2D(position, BASE_RADIUS).Any())
-                return true;
-            else if (zone.GetStaticUnits().OfType<Teleport>().WithinRange2D(position, PLAYER_RADIUS).Any())
-                return true;
-            else if (zone.PresenceManager.GetPresences().OfType<RandomSpawningExpiringPresence>().Where(p => p.SpawnOrigin.IsInRangeOf2D(position, BASE_RADIUS)).Any())
-                return true;
-
-            return zone.Players.WithinRange2D(position, range).Any();
-        }
-    }
-
-    public class GrowSpawnState : StaticSpawnState
-    {
-        private readonly GrowingPresence _growingPresence;
-        public GrowSpawnState(GrowingPresence presence, int playerMinDist = 200) : base(presence, playerMinDist)
-        {
-            _growingPresence = presence;
-        }
-
-        protected override void OnSpawned()
-        {
-            _presence.OnSpawned();
-            _presence.StackFSM.Push(new GrowthState(_growingPresence));
-        }
-    }
-
-    public class NPCBaseGrowState : GrowSpawnState
-    {
-        private readonly GrowingPresence _growingPresence;
-        public NPCBaseGrowState(GrowingPresence presence, int playerMinDist = 200) : base(presence, playerMinDist)
-        {
-            _growingPresence = presence;
-        }
-
-        protected override void SetSpawnDelay()
-        {
-            if (_growingPresence.CurrentGrowthLevel == 0)
-            {
-                _repawnDelayModifier = FastRandom.NextDouble(1.0, 2.0);
-            }
-            _delay = TimeSpan.FromSeconds((int)_presence.Configuration.DynamicLifeTime * _repawnDelayModifier);
-            _repawnDelayModifier = FastRandom.NextDouble(1.0, 2.0);
-        }
-
-        protected override bool IsValidSpawnPosition(Position position, int range)
-        {
-            var zone = _presence.Zone;
-            if (zone == null)
-            {
-                return false;
-            }
-
-            var radiusToCheck = 4;
-            var centerPosition = position;
-            for (var j = centerPosition.intY - radiusToCheck; j < centerPosition.intY + radiusToCheck; j++)
-            {
-                for (var i = centerPosition.intX - radiusToCheck; i < centerPosition.intX + radiusToCheck; i++)
-                {
-                    var cPos = new Position(i, j);
-                    if (!zone.Size.Contains(cPos.intX, cPos.intY))
-                        continue;
-
-                    var controlInfo = zone.Terrain.Controls.GetValue(position.intX, position.intY);
-                    if (controlInfo.IsAnyTerraformProtected || !zone.Terrain.Slope.CheckSlope(cPos.intX, cPos.intY, ZoneExtensions.MIN_SLOPE))
-                    {
-                        return false;
-                    }
-                }
-            }
-            return !IsInRange(position, range);
-        }
-
-        protected override Position FindSpawnPosition()
-        {
-            return _presence.PathFinder.FindSpawnPosition(_presence).ToPosition();
-        }
-    }
-
-    public class GrowthState : NullRoamingState
-    {
-        private readonly GrowingPresence _growingPresence;
-        private readonly TimeTracker _timer;
-        private int _currentLevel = 0;
-        public GrowthState(GrowingPresence presence) : base(presence)
-        {
-            _growingPresence = presence;
-            _timer = new TimeTracker(_growingPresence.GrowTime);
-            _currentLevel = _growingPresence.CurrentGrowthLevel;
-        }
-
-        public override void Update(TimeSpan time)
-        {
-            if (IsRunningTask)
-                return;
-
-            var members = GetAllMembers();
-            if (IsDeadAndExiting(members))
-                return;
-
-            if (!NextWaveReady(time))
-                return;
-
-            RunTask(() => SpawnNextWave(), t => { });
-        }
-
-        private bool NextWaveReady(TimeSpan time)
-        {
-            if (!CheckTimer(time))
-                return false;
-
-            _currentLevel++;
-            return true;
-        }
-
-        private bool CheckTimer(TimeSpan time)
-        {
-            _timer.Update(time);
-            if (!_timer.Expired)
-                return false;
-
-            _timer.Reset();
-            return true;
-        }
-
-        private void SpawnNextWave()
-        {
-            var flockConfigs = _growingPresence.Selector.GetFlocksForPresenceLevel(_growingPresence, _currentLevel);
-            foreach (var config in flockConfigs)
-            {
-                var flock = _growingPresence.CreateAndAddFlock(config);
-                flock.SpawnAllMembers();
-            }
-            _growingPresence.OnWaveSpawn(_currentLevel);
         }
     }
 }
